% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_auxillary.R
\name{find_mle_bfgs_log}
\alias{find_mle_bfgs_log}
\title{Function to find MLE via BFGS for Logistic Model}
\usage{
find_mle_bfgs_log(design, outcome)
}
\description{
Function to find MLE via BFGS for Logistic Model
}
