% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_auxillary.R
\name{find_mle_pseudoinverse}
\alias{find_mle_pseudoinverse}
\title{Function to find MLE via Pseudo inverse}
\usage{
find_mle_pseudoinverse(design, outcome)
}
\description{
Function to find MLE via Pseudo inverse
}
