% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_auxillary.R
\name{neg_gradient}
\alias{neg_gradient}
\title{we define the function neg_gradient which is the negative of the gradient.}
\usage{
neg_gradient(beta, design = design, outcome = outcome, noise_var = 1)
}
\description{
we define the function neg_gradient which is the negative of the gradient.
}
