% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_auxillary.R
\name{neg_log_likelihood}
\alias{neg_log_likelihood}
\title{we define the function neg_log_likelihood which is the negative of the log-likelihood.}
\usage{
neg_log_likelihood(beta, design = design, outcome = outcome, noise_var = 1)
}
\description{
we define the function neg_log_likelihood which is the negative of the log-likelihood.
}
