% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_auxillary.R
\name{newton_algo}
\alias{newton_algo}
\title{Function to find MLE via Newton's algorithm for Logistic Model}
\usage{
newton_algo(design, outcome)
}
\description{
Function to find MLE via Newton's algorithm for Logistic Model
}
