% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_auxillary.R
\name{newton_logistic}
\alias{newton_logistic}
\title{Function to perform Newton's method for logistic regression}
\usage{
newton_logistic(design, outcome, start_values = NULL, max_iter = 10000)
}
\description{
Function to perform Newton's method for logistic regression
}
