% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_logit.R
\name{hiper_glm}
\alias{hiper_glm}
\title{High-Performance GLM}
\usage{
hiper_glm(design, outcome, model = "linear", option = list())
}
\arguments{
\item{design}{A design matrix.}

\item{outcome}{The outcome variable.}

\item{model}{The type of model. Currently supports 'linear'.}

\item{option}{A list of options. Currently supports 'mle_solver' with values 'LS' or 'BFGS'.}
}
\value{
An object of class 'hiper_glm'.
}
\description{
Fits a high-performance generalized linear model.
}
\examples{
n_obs <- 10; n_pred <- 5
data <- simulate_data(n_obs, n_pred, model = "linear", seed = 1)
design <- data$design; outcome <- data$outcome
hiper_glm(design, outcome, model = "linear")
hiper_glm(design, outcome, model = "linear", option = list(mle_solver = "BFGS"))

}
