% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_logit.R
\name{hiper_glm}
\alias{hiper_glm}
\title{High-Performance GLM}
\usage{
hiper_glm(design, outcome)
}
\arguments{
\item{design}{A design matrix.}

\item{outcome}{The outcome variable.}
}
\value{
An object of class 'hiper_glm'.
}
\description{
Fits a high-performance generalized linear model.
}
